Separated classes and general utils from the gui_pack.py script. Have working subprocess logic for parse tab.
